fix tests

+++ src/test/kotlin/org/smolang/robust/domainSpecific/reasoner/ElGenerationTimeAnalyzerTest.kt
@@ -14,7 +14,8 @@ class ElGenerationTimeAnalyzerTest : StringSpec() {
             val mutationCounts = listOf(1,10,100)
             val input = File("src/test/resources/reasoners/ore_ont_155.owl")
             val output = File("src/test/resources/tempOutputs/timeData.csv")
-            output.delete()     // delete output file to get real result
+            if (output.exists())
+                output.delete()     // delete output file to get real result
             val timeout = 10000L // time in ms --> 10s
             ElGenerationTimeAnalyzer().timePerMutation(input, output, sampleSize, mutationCounts, timeout)
